Connected Authentication Backend with Front End Forms

+++ quiz-api/src/main/java/com/codersquiz/quiz_api/repositories/UserRepository.java
@@ -12,5 +12,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
 
-    Optional<User> findById(Integer userId);
+    //Optional<User> findById(Integer userId);
+    Optional<User> findById(Long userId); // This is not needed as JpaRepository already provides this.
 }
